fix: pequeño refactor scoringNotSurpassesThreshold
